# core/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/app/directory/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "平台创建数据目录"
                ],
                "summary": "平台创建数据目录",
                "parameters": [
                    {
                        "description": "平台创建数据目录",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.InternalDirectoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "平台创建数据目录",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.InternalDirectoryRequest": {
            "type": "object",
            "properties": {
                "direct_id": {
                    "type": "integer"
                },
                "directory_name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "remark": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}